## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов

Результаты бенчмарка:
old:
    BenchmarkGetDomainStat-10             14      78174158 ns/op
    stats_optimization_test.go:45: time used: 356.119333ms / 300ms
    stats_optimization_test.go:46: memory used: 307Mb / 30Mb
new:
    BenchmarkGetDomainStat-10          19369         61512 ns/op
    stats_optimization_test.go:45: time used: 146.379ms / 300ms
    stats_optimization_test.go:46: memory used: 13Mb / 30Mb
    stats_optimization_test.go:45: time used: 144.517542ms / 300ms
    stats_optimization_test.go:46: memory used: 13Mb / 30Mb
    stats_optimization_test.go:45: time used: 144.649875ms / 300ms
    stats_optimization_test.go:46: memory used: 13Mb / 30Mb
    stats_optimization_test.go:45: time used: 146.741792ms / 300ms
    stats_optimization_test.go:46: memory used: 13Mb / 30Mb
    stats_optimization_test.go:45: time used: 144.512957ms / 300ms
    stats_optimization_test.go:46: memory used: 13Mb / 30Mb